Fix generation of UV map
The previous SMPL UV map directly mapped each
individual texture index to a vertex corner; however,
there is no guarantee that the OBJ file would adhere
to this mapping.

This fix changes this so that we map each tuple of
texture indices (e.g. each triangle) to its corresponding
vertex triangle.

This commit also includes a method to render the
resulting vertex mapping to a point cloud for easier
debugging.